App adjustments and Search Templates action
